Replace old SortableTable component
