+++ src/championOverview/ChampionOverview.tsx
@@ -2,7 +2,7 @@ import { ColumnDef, createColumnHelper } from "@tanstack/react-table";
 import React from "react";
 import { useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
-import { SortableTable } from "../components/SortableTable2";
+import { SortableTable } from "../components/SortableTable";
 import { processChampions } from "../logic/statsProcessors";
 import { statsSelector } from "../redux/statsSelectors";
 import { Champion } from "../types/domain/Champion";
