+++ src/playerOverview/PlayerScreen.tsx
@@ -2,7 +2,7 @@ import { ColumnDef, createColumnHelper } from "@tanstack/react-table";
 import React from "react";
 import { useSelector } from "react-redux";
 import { useLoaderData, useNavigate } from "react-router-dom";
-import { SortableTable } from "../components/SortableTable2";
+import { SortableTable } from "../components/SortableTable";
 import { processPlayerChampions } from "../logic/statsProcessors";
 import { AppState } from "../redux/rootReducer";
 import { statsSelector } from "../redux/statsSelectors";
